+++ server/src/suggestions/builtins/index.ts
@@ -1,9 +1,2 @@
-export { BUILTIN_ENTRIES } from './catalog';
 export { collectBuiltinSuggestions } from './filter';
 export { collectBuiltinFunctionSuggestions } from './signatures';
-export {
-  BUILTIN_CATEGORIES,
-  builtinEntryToSuggestion,
-  type BuiltinCategory,
-  type BuiltinEntry,
-} from './types';